+++ mlir/lib/Transform/CMakeLists.txt
@@ -2,6 +2,8 @@ add_mlir_library(MLIRLapisTransforms
   AlgebraicKernelFusion.cpp
   ContractionPlanner.cpp
   Einsum.cpp
+  FuseAlgebraicKernelLoops.cpp
+  OutlineAlgebraicKernels.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LAPIS_SOURCE_DIR}/mlir/include/lapis/Transform
